src/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
defusedxml==0.6.0
Django==3.0.5
django-rest-multiple-models==2.1.3
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
pytz==2019.3
sqlparse==0.3.1
wincertstore==0.2
